Added some tests, making it a but more robust

+++ uninstall.php
@@ -4,8 +4,9 @@
 if (!defined('WP_UNINSTALL_PLUGIN')) {
     die;
 }
+
+if ( ! current_user_can( 'activate_plugins' ) )
+       return;
  
 $option_name = 'slug-length-extender-wpver';
- 
 delete_option($option_name);
-
